refactor(edge/watchdog): split watchdog into config coordinator and time module
